Update
Solves #14 and #24.
Changed EG_KITCHEN to KITCHEN.
Added a research for Kitchen.
