+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>me.mrCookieSlime</groupId>
     <artifactId>ExoticGarden</artifactId>
-    <version>1.6.3</version>
+    <version>1.7.0</version>
     <properties>
 	    <maven.compiler.source>1.8</maven.compiler.source>
 	    <maven.compiler.target>1.8</maven.compiler.target>
@@ -60,12 +60,12 @@
         <dependency>
 	        <groupId>com.github.TheBusyBiscuit</groupId>
 	        <artifactId>CS-CoreLib</artifactId>
-	        <version>v1.5.16</version>
+	        <version>v1.5.17</version>
 	    </dependency>
         <dependency>
 	        <groupId>com.github.TheBusyBiscuit</groupId>
 	        <artifactId>Slimefun4</artifactId>
-	        <version>v4.1.12</version>
+	        <version>v4.1.14</version>
 	    </dependency>
     </dependencies>
 </project>
