+++ src/me/mrCookieSlime/ExoticGarden/Schematic/org/jnbt/ByteArrayTag.java
@@ -1,4 +1,4 @@
-package org.jnbt;
+package me.mrCookieSlime.ExoticGarden.Schematic.org.jnbt;
 
 /*
  * JNBT License
@@ -34,40 +34,48 @@ package org.jnbt;
  */
 
 /**
- * The <code>TAG_Float</code> tag.
+ * The <code>TAG_Byte_Array</code> tag.
  * @author Graham Edgecombe
  *
  */
-public final class FloatTag extends Tag {
-
+public final class ByteArrayTag extends Tag {
+	
 	/**
 	 * The value.
 	 */
-	private final float value;
+	private final byte[] value;
 	
 	/**
 	 * Creates the tag.
 	 * @param name The name.
 	 * @param value The value.
 	 */
-	public FloatTag(String name, float value) {
+	public ByteArrayTag(String name, byte[] value) {
 		super(name);
 		this.value = value;
 	}
 	
 	@Override
-	public Float getValue() {
+	public byte[] getValue() {
 		return value;
 	}
 	
 	@Override
 	public String toString() {
+		StringBuilder hex = new StringBuilder();
+		for(byte b : value) {
+			String hexDigits = Integer.toHexString(b).toUpperCase();
+			if(hexDigits.length() == 1) {
+				hex.append("0");
+			}
+			hex.append(hexDigits).append(" ");
+		}
 		String name = getName();
 		String append = "";
 		if(name != null && !name.equals("")) {
 			append = "(\"" + this.getName() + "\")";
 		}
-		return "TAG_Float" + append + ": " + value;
+		return "TAG_Byte_Array" + append + ": " + hex.toString();
 	}
 
 }
