+++ src/me/mrCookieSlime/ExoticGarden/Schematic/org/jnbt/CompoundTag.java
@@ -1,4 +1,4 @@
-package org.jnbt;
+package me.mrCookieSlime.ExoticGarden.Schematic.org.jnbt;
 
 /*
  * JNBT License
@@ -33,28 +33,50 @@ package org.jnbt;
  * POSSIBILITY OF SUCH DAMAGE. 
  */
 
+import java.util.Collections;
+import java.util.Map;
+
 /**
- * The <code>TAG_End</code> tag.
+ * The <code>TAG_Compound</code> tag.
  * @author Graham Edgecombe
  *
  */
-public final class EndTag extends Tag {
-
+public final class CompoundTag extends Tag {
+	
+	/**
+	 * The value.
+	 */
+	private final Map<String, Tag> value;
+	
 	/**
 	 * Creates the tag.
+	 * @param name The name.
+	 * @param value The value.
 	 */
-	public EndTag() {
-		super("");
+	public CompoundTag(String name, Map<String, Tag> value) {
+		super(name);
+		this.value = Collections.unmodifiableMap(value);
 	}
 
 	@Override
-	public Object getValue() {
-		return null;
+	public Map<String, Tag> getValue() {
+		return value;
 	}
 	
 	@Override
 	public String toString() {
-		return "TAG_End";
+		String name = getName();
+		String append = "";
+		if(name != null && !name.equals("")) {
+			append = "(\"" + this.getName() + "\")";
+		}
+		StringBuilder bldr = new StringBuilder();
+		bldr.append("TAG_Compound" + append + ": " + value.size() + " entries\r\n{\r\n");
+		for(Map.Entry<String, Tag> entry : value.entrySet()) {
+			bldr.append("   " + entry.getValue().toString().replaceAll("\r\n", "\r\n   ") + "\r\n");
+		}
+		bldr.append("}");
+		return bldr.toString();
 	}
 
 }
