+++ src/me/mrCookieSlime/ExoticGarden/Schematic/org/jnbt/DoubleTag.java
@@ -1,4 +1,4 @@
-package org.jnbt;
+package me.mrCookieSlime.ExoticGarden.Schematic.org.jnbt;
 
 /*
  * JNBT License
@@ -34,37 +34,40 @@ package org.jnbt;
  */
 
 /**
- * Represents a single NBT tag.
+ * The <code>TAG_Double</code> tag.
  * @author Graham Edgecombe
  *
  */
-public abstract class Tag {
+public final class DoubleTag extends Tag {
 	
 	/**
-	 * The name of this tag.
+	 * The value.
 	 */
-	private final String name;
-	
+	private final double value;
+
 	/**
-	 * Creates the tag with the specified name.
+	 * Creates the tag.
 	 * @param name The name.
+	 * @param value The value.
 	 */
-	public Tag(String name) {
-		this.name = name;
+	public DoubleTag(String name, double value) {
+		super(name);
+		this.value = value;
 	}
 	
-	/**
-	 * Gets the name of this tag.
-	 * @return The name of this tag.
-	 */
-	public final String getName() {
-		return name;
+	@Override
+	public Double getValue() {
+		return value;
 	}
 	
-	/**
-	 * Gets the value of this tag.
-	 * @return The value of this tag.
-	 */
-	public abstract Object getValue();
+	@Override
+	public String toString() {
+		String name = getName();
+		String append = "";
+		if(name != null && !name.equals("")) {
+			append = "(\"" + this.getName() + "\")";
+		}
+		return "TAG_Double" + append + ": " + value;
+	}
 
 }
