+++ src/me/mrCookieSlime/ExoticGarden/Schematic/org/jnbt/EndTag.java
@@ -1,4 +1,4 @@
-package org.jnbt;
+package me.mrCookieSlime.ExoticGarden.Schematic.org.jnbt;
 
 /*
  * JNBT License
@@ -34,48 +34,27 @@ package org.jnbt;
  */
 
 /**
- * The <code>TAG_Byte_Array</code> tag.
+ * The <code>TAG_End</code> tag.
  * @author Graham Edgecombe
  *
  */
-public final class ByteArrayTag extends Tag {
-	
-	/**
-	 * The value.
-	 */
-	private final byte[] value;
-	
+public final class EndTag extends Tag {
+
 	/**
 	 * Creates the tag.
-	 * @param name The name.
-	 * @param value The value.
 	 */
-	public ByteArrayTag(String name, byte[] value) {
-		super(name);
-		this.value = value;
+	public EndTag() {
+		super("");
 	}
-	
+
 	@Override
-	public byte[] getValue() {
-		return value;
+	public Object getValue() {
+		return null;
 	}
 	
 	@Override
 	public String toString() {
-		StringBuilder hex = new StringBuilder();
-		for(byte b : value) {
-			String hexDigits = Integer.toHexString(b).toUpperCase();
-			if(hexDigits.length() == 1) {
-				hex.append("0");
-			}
-			hex.append(hexDigits).append(" ");
-		}
-		String name = getName();
-		String append = "";
-		if(name != null && !name.equals("")) {
-			append = "(\"" + this.getName() + "\")";
-		}
-		return "TAG_Byte_Array" + append + ": " + hex.toString();
+		return "TAG_End";
 	}
 
 }
