+++ src/me/mrCookieSlime/ExoticGarden/Schematic/org/jnbt/NBTInputStream.java
@@ -1,4 +1,4 @@
-package org.jnbt;
+package me.mrCookieSlime.ExoticGarden.Schematic.org.jnbt;
 
 /*
  * JNBT License
