+++ src/me/mrCookieSlime/ExoticGarden/Schematic/org/jnbt/NBTOutputStream.java
@@ -1,4 +1,4 @@
-package org.jnbt;
+package me.mrCookieSlime.ExoticGarden.Schematic.org.jnbt;
 
 import java.io.Closeable;
 import java.io.DataOutputStream;
