+++ src/me/mrCookieSlime/ExoticGarden/Schematic/org/jnbt/Tag.java
@@ -1,4 +1,4 @@
-package org.jnbt;
+package me.mrCookieSlime.ExoticGarden.Schematic.org.jnbt;
 
 /*
  * JNBT License
@@ -34,40 +34,37 @@ package org.jnbt;
  */
 
 /**
- * The <code>TAG_String</code> tag.
+ * Represents a single NBT tag.
  * @author Graham Edgecombe
  *
  */
-public final class StringTag extends Tag {
-
-	/**
-	 * The value.
-	 */
-	private final String value;
+public abstract class Tag {
 	
 	/**
-	 * Creates the tag.
+	 * The name of this tag.
+	 */
+	private final String name;
+	
+	/**
+	 * Creates the tag with the specified name.
 	 * @param name The name.
-	 * @param value The value.
 	 */
-	public StringTag(String name, String value) {
-		super(name);
-		this.value = value;
+	public Tag(String name) {
+		this.name = name;
 	}
 	
-	@Override
-	public String getValue() {
-		return value;
+	/**
+	 * Gets the name of this tag.
+	 * @return The name of this tag.
+	 */
+	public final String getName() {
+		return name;
 	}
 	
-	@Override
-	public String toString() {
-		String name = getName();
-		String append = "";
-		if(name != null && !name.equals("")) {
-			append = "(\"" + this.getName() + "\")";
-		}
-		return "TAG_String" + append + ": " + value;
-	}
+	/**
+	 * Gets the value of this tag.
+	 * @return The value of this tag.
+	 */
+	public abstract Object getValue();
 
 }
