+++ src/me/mrCookieSlime/ExoticGarden/Tree.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
 
+import me.mrCookieSlime.ExoticGarden.Schematic.Schematic;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 
 import org.bukkit.Material;
