+++ src/plugin.yml
@@ -1,6 +1,6 @@
 name: ExoticGarden
 version: ${project.version}
-authors: [mrCookieSlime, Goldentoenail, Joapple]
+authors: [mrCookieSlime, Goldentoenail, Joapple, SoSeDiK]
 description: A Slimefun Addon that adds new Plants and dishes to the game
 website: http://TheBusyBiscuit.github.io/
